refactor the code to remove redundant formatLocalIsoFromString function
