+++ manager/static/manager/utils.js
@@ -1,5 +1,6 @@
-function formatLocalISOFromString(dateUTCString) {
-  const dateTime = new Date(dateUTCString);
+function formatLocalISO(dateUTC) {
+  // formats UTC time to local time
+  const dateTime = new Date(dateUTC);
   const timeZoneOffSet = dateTime.getTimezoneOffset() * 60 * 1000;
   let dateTimeLocal = dateTime - timeZoneOffSet;
   dateTimeLocal = new Date(dateTimeLocal);
@@ -9,4 +10,23 @@ function formatLocalISOFromString(dateUTCString) {
   return iso;
 }
 
-export {formatLocalISOFromString}
+function getValidDateISOString(dateLocalStr) {
+  // formats local time to UTC time.
+  // returns ISO string of that date, today midnight if no values are provided
+  if(dateLocalStr !== "") {
+    return new Date(dateLocalStr).toISOString()
+  }
+  let defaultDate =  new Date()
+  defaultDate.setHours(0,0,0,0)
+  defaultDate.setDate(defaultDate.getDate() + 1)
+  return defaultDate.toISOString()
+}
+
+function getValidEstimatedTime(time) {
+  if(time === "") {
+    return 0;
+  }
+  return Number(time);
+}
+
+export {formatLocalISO, getValidDateISOString, getValidEstimatedTime}
